+++ docs/api/changelog.rst
@@ -8,6 +8,8 @@ Changelog
 0.6.4 - TBD
     * Improved error messages when using ``Concrete.cast_as_concrete`` and ``Concrete.type_var``
     * Ensure that ``Concrete.type_var`` is only used in the module scope
+    * Fixed a bug where it's possible for virtual dependency cleanup to try remove a file
+      that doesn't exist
 
 .. _release-0.6.3: